fix normal mapping

+++ BasicRenderer/include/mesh.h
@@ -16,7 +16,7 @@ struct GeometryData {
     std::vector<float> normals;
     std::vector<uint32_t> indices;
     std::vector<float> texcoords;
-    std::vector<float> joints;
+    std::vector<UINT> joints;
     std::vector<float> weights;
     std::shared_ptr<Material> material;
 };
